v2.4.0-beta.1. remove circleci, cleaning in ws client
