+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "bybit-api",
-  "version": "2.3.2",
+  "version": "2.4.0-beta.1",
   "description": "Node.js connector for Bybit's REST APIs and WebSockets, with TypeScript & integration tests.",
   "main": "lib/index.js",
   "types": "lib/index.d.ts",
