+++ src/types/index.ts
@@ -1,3 +1,4 @@
 export * from './response';
 export * from './request';
 export * from './shared';
+export * from './websockets';
