+++ src/util/index.ts
@@ -2,3 +2,4 @@ export * from './BaseRestClient';
 export * from './requestUtils';
 export * from './WsStore';
 export * from './logger';
+export * from './websocket-util';
